Add files via upload
The Food Recipe Book is a delightful web application designed to bring culinary inspiration to your fingertips. Using the Spoonacular API, this application showcases a curated collection of random recipes, complete with enticing images, ingredient lists, and a sleek, responsive design. With a modern touch, the 3D effect on recipe view buttons enhances the user experience. The simplicity of the codebase, implemented in HTML, CSS, and JavaScript, makes it easy to explore, customize, and build upon. Whether you're a food enthusiast or a developer looking for a stylish template, the Food Recipe Book invites you to embark on a culinary journey. Clone the repository, personalize it to your taste, and enjoy the fusion of technology and gastronomy!